# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Candlekeep Core",
    "description": "FastAPI backend for RP platform with character cards and chat management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "List Providers",
        "description": "List configured model providers",
        "operationId": "list_providers_api_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderResponse"
                  },
                  "type": "array",
                  "title": "Response List Providers Api Providers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Create Provider",
        "description": "Register a new model provider",
        "operationId": "create_provider_api_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{provider_id}": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Get Provider",
        "description": "Get provider details by ID",
        "operationId": "get_provider_api_providers__provider_id__get",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "providers"
        ],
        "summary": "Update Provider",
        "description": "Update provider configuration",
        "operationId": "update_provider_api_providers__provider_id__put",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{provider_id}/flags": {
      "patch": {
        "tags": [
          "providers"
        ],
        "summary": "Update Provider Flags",
        "description": "Enable or disable a provider",
        "operationId": "update_provider_flags_api_providers__provider_id__flags_patch",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderFlagsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProviderCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name",
            "description": "Unique provider name"
          },
          "provider_type": {
            "$ref": "#/components/schemas/ProviderType",
            "description": "Type of provider"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url",
            "description": "API endpoint URL (optional, uses default for known providers)"
          },
          "api_key_env_var": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env Var",
            "description": "Environment variable name for API key (required for CUSTOM providers)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "provider_type"
        ],
        "title": "ProviderCreate",
        "description": "Schema for creating a new provider"
      },
      "ProviderFlagsUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "ProviderFlagsUpdate",
        "description": "Schema for updating provider flags only"
      },
      "ProviderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "provider_type": {
            "type": "string",
            "title": "Provider Type"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "api_key_configured": {
            "type": "boolean",
            "title": "Api Key Configured",
            "description": "Whether API key is available in environment"
          },
          "env_var_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Env Var Name",
            "description": "Expected environment variable name for API key"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "provider_type",
          "base_url",
          "enabled",
          "created_at",
          "updated_at",
          "api_key_configured",
          "env_var_name"
        ],
        "title": "ProviderResponse",
        "description": "Schema for provider responses"
      },
      "ProviderType": {
        "type": "string",
        "enum": [
          "xai",
          "google",
          "openai",
          "anthropic",
          "openrouter",
          "ollama",
          "custom"
        ],
        "title": "ProviderType",
        "description": "Supported provider types"
      },
      "ProviderUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "api_key_env_var": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env Var",
            "description": "Environment variable name for API key (CUSTOM providers only)"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "type": "object",
        "title": "ProviderUpdate",
        "description": "Schema for updating a provider"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}